Split classes in different files.
Separated classes by their themes in individual files.
