+++ day_widget.py
@@ -10,10 +10,9 @@ a sidebar for managing flex time and vacation requests, and information panels
 to display the user's current work status.
 """
 import tkinter as tk
-import calendar
 
 
-class DayWidget(tk.Button):
+class DayWidget(tk.Frame):
     """
     A class used to represent a working day element in the
     SCT time management GUI.
@@ -212,173 +211,3 @@ class DayWidget(tk.Button):
             time_string = f"{hours:02}:{minutes:02}"
 
         return time_string
-
-
-class CalendarHeader(tk.Frame):
-    """
-    A header frame for the calendar widget, containing navigation buttons.
-
-    Methods
-    -------
-    __init__(parent)
-        Initializes the header with navigation buttons
-        and a label for the month.
-    """
-
-    def __init__(self, parent):
-        """
-        Constructs the CalendarHeader, including navigation
-        and display elements.
-
-        Parameters
-        ----------
-        parent : widget
-            The parent widget in which this header widget will be placed.
-        """
-        super().__init__(master=parent, relief="solid", borderwidth=1)
-
-        tk.Button(self, text="<").grid(column=0, row=0)
-        tk.Button(self, text=">").grid(column=2, row=0)
-
-        self.var_selected_month = tk.StringVar(value="Month YYYY")
-        tk.Label(self, textvariable=self.var_selected_month).grid(column=1,
-                                                                  row=0)
-
-
-class CalendarContent(tk.Frame):
-    """
-    The main content area of the calendar widget, containing day widgets.
-
-    Methods
-    -------
-    __init__(parent)
-        Initializes the calendar grid with day widgets.
-    """
-
-    def __init__(self, parent):
-        """
-        Constructs the CalendarContent with day labels
-        and DayWidget instances.
-
-        Parameters
-        ----------
-        parent : widget
-            The parent widget in which this calendar
-            content widget will be placed.
-        """
-
-        super().__init__(master=parent, relief="solid", borderwidth=1, padx=10,
-                         pady=10)
-
-        self.days = [0] * 35
-
-        for i in range(0, 7):
-            day = calendar.day_name[i]
-            tk.Label(self, text=day).grid(row=0, column=i)
-
-        for i in range(0, 35):
-            self.days[i] = DayWidget(self)
-            day = self.days[i]
-            day.grid(row=i // 7 + 1, column=i % 7, padx=5, pady=5)
-            day.set_day_number(i)
-
-
-class CalendarWidget(tk.Frame):
-    """
-    A calendar widget containing a header and content with day elements.
-
-    Methods
-    -------
-    __init__(parent)
-        Initializes the CalendarWidget with header and content.
-    """
-
-    def __init__(self, parent):
-        """
-        Constructs all necessary attributes for the CalendarWidget object.
-
-        Parameters
-        ----------
-        parent : widget
-            The parent widget in which this calendar widget will be placed.
-        """
-        super().__init__(master=parent, relief="solid", borderwidth=1)
-
-        self.header = CalendarHeader(self)
-        self.content = CalendarContent(self)
-
-        self.header.pack(expand=True, fill=tk.BOTH)
-        self.content.pack()
-
-
-class SideBar(tk.Frame):
-    """
-    A sidebar component containing action buttons and an information panel.
-
-    Methods
-    -------
-    __init__(parent)
-        Initializes the sidebar with buttons and an information panel.
-    """
-
-    def __init__(self, parent):
-        """
-        Constructs the SideBar with vacation and time logging buttons.
-
-        Parameters
-        ----------
-        parent : widget
-            The parent widget in which this sidebar will be placed.
-        """
-
-        super().__init__(master=parent, bg="azure3", padx=20, pady=20)
-
-        self.info_panel = InfoPanel(self)
-
-        button_request_vacation = tk.Button(self, text="Request Vacation")
-        button_log_working_time = tk.Button(self, text="Start Workday")
-        button_log_break_time = tk.Button(self, text="Start BreaK")
-
-        self.info_panel.pack(fill=tk.X, pady=5)
-        button_request_vacation.pack(side="top", fill=tk.X, pady=5)
-        button_log_break_time.pack(side="bottom", fill=tk.X, pady=5)
-        button_log_working_time.pack(side="bottom", fill=tk.X, pady=5)
-
-
-class InfoPanel(tk.Frame):
-    """
-    A panel for displaying user information such as
-    flex time and vacation days.
-
-    Methods
-    -------
-    __init__(parent)
-        Initializes the information panel with labels for user status.
-    """
-
-    def __init__(self, parent):
-        """
-        Constructs the InfoPanel to display flex time and vacation details.
-
-        Parameters
-        ----------
-        parent : widget
-            The parent widget in which this information panel will be placed.
-        """
-
-        super().__init__(master=parent, padx=10, pady=10,
-                         borderwidth=1, relief="solid")
-
-        self.var_flex_time = tk.StringVar(value="+ 27:13")
-        self.var_vacation_days = tk.StringVar(value=17)
-        self.var_old_vacation_days = tk.StringVar(value=0)
-
-        tk.Label(self, text="Flex-Time:").pack(anchor="nw")
-        tk.Label(self, textvariable=self.var_flex_time).pack(anchor="nw")
-
-        tk.Label(self, text="Remaining vacation days:").pack(anchor="nw")
-        tk.Label(self, textvariable=self.var_vacation_days).pack(anchor="nw")
-
-        tk.Label(self, text="Vacation from last year:").pack(anchor="nw")
-        tk.Label(
-            self, textvariable=self.var_old_vacation_days).pack(anchor="nw")
